feat!: wait for watcher to be ready
Patch the "ready" event of chokidar so it‘s emitted more than once. This allows us to reliably postpone the promise resolution (and "finish" event) of each generator run until the chokidar is ready. This is particularly useful for scenarios where you await the generator run and immediately write/unlink a file afterwards (e.g. for integration tests).

I‘ve also updated the public API of generator instances. A new `watcher` property has been added, which is an object that replaces the `isWatchMode` property and contains the old `watchedFiles` and `blamedFiles` properties.